requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.11.5
django-paypal==0.5.0
djangorestframework==3.7.7
idna==2.6
paypal==1.2.5
Pillow==5.0.0
pytz==2018.3
requests==2.18.4
simplejson==3.13.2
six==1.11.0
urllib3==1.22
